Fix windows abstract, added minimum and maximum sizes for objects

+++ GUI/src/ru/konstanteam/lokutils/gui/objects/margin/Margin.java
@@ -1,33 +1,33 @@
 package ru.konstanteam.lokutils.gui.objects.margin;
 
 public class Margin {
-    public static final Margin PIXEL_MARGIN = new Margin(1,1,1,1);
+    public static final Margin PIXEL_MARGIN = new Margin(1, 1, 1, 1);
 
-    protected final int left;
-    protected final int right;
-    protected final int top;
-    protected final int bottom;
+    protected final float left;
+    protected final float right;
+    protected final float top;
+    protected final float bottom;
 
-    public Margin(int left, int right, int top, int bottom) {
+    public Margin(float left, float right, float top, float bottom) {
         this.left = left;
         this.right = right;
         this.top = top;
         this.bottom = bottom;
     }
 
-    public int getLeft() {
+    public float getLeft() {
         return left;
     }
 
-    public int getRight() {
+    public float getRight() {
         return right;
     }
 
-    public int getTop() {
+    public float getTop() {
         return top;
     }
 
-    public int getBottom() {
+    public float getBottom() {
         return bottom;
     }
 }
